backend/similar/backend_api/requirements.txt
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
async-generator==1.10
attrs==23.1.0
autopep8==2.0.2
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
executing==1.2.0
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.23.1
ipython==8.12.2
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.3
outcome==1.2.0
packaging==23.1
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg==3.1.9
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.15.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.1.0
random-password-generator==2.2.0
requests==2.28.2
selenium==4.9.1
six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-103/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
tornado==6.3.2
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.6.2
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
wsproto==1.2.0
zipp==3.15.0
